Refactoring codespace.
