+++ .gitignore
@@ -1,7 +1,7 @@
 .venv/
 .env*
 venv/
-env/
+.env/
 
 __pycache__/
 *.py[cod]
