+++ backend/task_manager/src/main.py
@@ -1,11 +1,11 @@
 '''
-Main file of projects.
+Main file of cohort_manager.
 '''
 
 
 from fastapi import FastAPI
 
-from MathLab.core.database import lifespan
+from utils.database import lifespan
 
 app = FastAPI(lifespan=lifespan)
