--- requirements.txt
@@ -1,22 +1,0 @@
-alembic==1.14.0
-annotated-types==0.7.0
-anyio==4.6.2.post1
-asyncpg==0.30.0
-click==8.1.7
-colorama==0.4.6
-fastapi==0.115.5
-greenlet==3.1.1
-h11==0.14.0
-idna==3.10
-Mako==1.3.6
-MarkupSafe==3.0.2
-psycopg2==2.9.10
-pydantic==2.9.2
-pydantic-settings==2.6.1
-pydantic_core==2.23.4
-python-dotenv==1.0.1
-sniffio==1.3.1
-SQLAlchemy==2.0.36
-starlette==0.41.2
-typing_extensions==4.12.2
-uvicorn==0.32.0
